Upgraded with tweaks made during the 2016 Intro to Linux course.
I think the main addition is the use of video, some extra handling in
the dual-column layout to pick code fragments for the righthand-side
column if there is no image on the slide, and support for inlining
shell commands and citations.

A bigger change is the javascript support for aspect ratios. This used
to be server-side on an old version, but now works with only client-side
CSS. Two index files are created for the slide show index-169.html for
16:9 ratio slides, and index-43.html for 4:3.

Shortly after this I discovered that our ugly VGA projectors are actually
8:10, which is weird and unusual but life is too short...